refactor: fix capitals in filename
